Ubuntu 11.04 build

+++ src/lib/rt/src/rt_mh_outunit.c
@@ -420,7 +420,7 @@ mh_OutunitReceive ()
     return MH__QCOMGETMSG;
   }
 
-  switch (msg.type.b) {
+  switch ((int)msg.type.b) {
 
   case qcom_eBtype_qcom:
     fromQcom(&msg);
